# Fix incomplete hash key in sequence table data.