--- a/src/backend/commands/sequence.c
+++ b/src/backend/commands/sequence.c
@@ -77,6 +77,16 @@ typedef struct sequence_magic
 	uint32		magic;
 } sequence_magic;
 
+typedef struct SeqTableKey
+{
+	Oid relid;						/* pg_class OID of this sequence */
+	bool called_from_dispatcher;	/* sequence called from dispatcher */
+}
+#if defined(pg_attribute_packed)
+			pg_attribute_packed()
+#endif
+SeqTableKey;
+
 /*
  * We store a SeqTable item for every sequence we have touched in the current
  * session.  This is needed to hold onto nextval/currval state.  (We can't
@@ -85,7 +95,7 @@ typedef struct sequence_magic
  */
 typedef struct SeqTableData
 {
-	Oid			relid;			/* pg_class OID of this sequence (hash key) */
+	SeqTableKey	key;			/* sequence data hash key */
 	Oid			filenode;		/* last seen relfilenode of this sequence */
 	LocalTransactionId lxid;	/* xact in which we last did a seq op */
 	bool		last_valid;		/* do we have a valid "last" value? */
@@ -99,16 +109,6 @@ typedef struct SeqTableData
 typedef SeqTableData *SeqTable;
 
 static HTAB *seqhashtab = NULL; /* hash table for SeqTable items */
-
-typedef struct SeqTableKey
-{
-	Oid relid;
-	bool called_from_dispatcher;
-}
-#if defined(pg_attribute_packed)
-			pg_attribute_packed()
-#endif
-SeqTableKey;
 
 /*
  * last_used_seq is updated by nextval() to point to the last used
@@ -707,7 +707,7 @@ nextval_internal(Oid relid, bool check_permissions, bool called_from_dispatcher)
 	init_sequence_internal(relid, &elm, &seqrel, called_from_dispatcher);
 
 	if (check_permissions &&
-		pg_class_aclcheck(elm->relid, GetUserId(),
+		pg_class_aclcheck(elm->key.relid, GetUserId(),
 						  ACL_USAGE | ACL_UPDATE) != ACLCHECK_OK)
 		ereport(ERROR,
 				(errcode(ERRCODE_INSUFFICIENT_PRIVILEGE),
@@ -965,7 +965,7 @@ currval_oid(PG_FUNCTION_ARGS)
 	/* open and lock sequence */
 	init_sequence(relid, &elm, &seqrel);
 
-	if (pg_class_aclcheck(elm->relid, GetUserId(),
+	if (pg_class_aclcheck(elm->key.relid, GetUserId(),
 						  ACL_SELECT | ACL_USAGE) != ACLCHECK_OK)
 		ereport(ERROR,
 				(errcode(ERRCODE_INSUFFICIENT_PRIVILEGE),
@@ -1005,7 +1005,7 @@ lastval(PG_FUNCTION_ARGS)
 				 errmsg("lastval is not yet defined in this session")));
 
 	/* Someone may have dropped the sequence since the last nextval() */
-	if (!SearchSysCacheExists1(RELOID, ObjectIdGetDatum(last_used_seq->relid)))
+	if (!SearchSysCacheExists1(RELOID, ObjectIdGetDatum(last_used_seq->key.relid)))
 		ereport(ERROR,
 				(errcode(ERRCODE_OBJECT_NOT_IN_PREREQUISITE_STATE),
 				 errmsg("lastval is not yet defined in this session")));
@@ -1015,7 +1015,7 @@ lastval(PG_FUNCTION_ARGS)
 	/* nextval() must have already been called for this sequence */
 	Assert(last_used_seq->last_valid);
 
-	if (pg_class_aclcheck(last_used_seq->relid, GetUserId(),
+	if (pg_class_aclcheck(last_used_seq->key.relid, GetUserId(),
 						  ACL_SELECT | ACL_USAGE) != ACLCHECK_OK)
 		ereport(ERROR,
 				(errcode(ERRCODE_INSUFFICIENT_PRIVILEGE),
@@ -1064,7 +1064,7 @@ do_setval(Oid relid, int64 next, bool iscalled)
 	/* open and lock sequence */
 	init_sequence(relid, &elm, &seqrel);
 
-	if (pg_class_aclcheck(elm->relid, GetUserId(), ACL_UPDATE) != ACLCHECK_OK)
+	if (pg_class_aclcheck(elm->key.relid, GetUserId(), ACL_UPDATE) != ACLCHECK_OK)
 		ereport(ERROR,
 				(errcode(ERRCODE_INSUFFICIENT_PRIVILEGE),
 				 errmsg("permission denied for sequence %s",
@@ -1210,7 +1210,7 @@ lock_and_open_sequence(SeqTable seq)
 		currentOwner = CurrentResourceOwner;
 		CurrentResourceOwner = TopTransactionResourceOwner;
 
-		LockRelationOid(seq->relid, RowExclusiveLock);
+		LockRelationOid(seq->key.relid, RowExclusiveLock);
 
 		CurrentResourceOwner = currentOwner;
 
@@ -1219,7 +1219,7 @@ lock_and_open_sequence(SeqTable seq)
 	}
 
 	/* We now know we have the lock, and can safely open the rel */
-	return relation_open(seq->relid, NoLock);
+	return relation_open(seq->key.relid, NoLock);
 }
 
 /*