New sequence hash key, introduced in d874700, wasn't taken into account by `SeqTable` struct, which should contain the whole key to avoid key data corruption. `SeqTable` now using whole `SeqTableKey` struct to avoid future bugs related to key extension.

Issue #12583.