@kainwen 2021-09-15
Thanks @InnerLife0 Would you  please share some performance data here, or in that issue context to have a compare?