@InnerLife0 2021-09-15
@kainwen Here is some analysis for relatively small amount of data on 3-segment demo cluster.
```
create table seq_test(i serial, j integer) distributed by (i);

--before d874700
explain analyze insert into seq_test(j) select generate_series(1, 100000);
                                                            QUERY PLAN                                                             
-----------------------------------------------------------------------------------------------------------------------------------
 Insert on seq_test  (cost=0.00..0.02 rows=1 width=8) (never executed)
   ->  Result  (cost=0.00..0.00 rows=1 width=12) (actual time=80.786..7155.867 rows=33462 loops=1)
         ->  Redistribute Motion 1:3  (slice1)  (cost=0.00..0.00 rows=1 width=8) (actual time=80.782..7150.094 rows=33462 loops=1)
               Hash Key: (int4(nextval('seq_test_i_seq'::regclass)))
               ->  Result  (cost=0.00..0.00 rows=1 width=8) (actual time=0.264..7007.020 rows=100000 loops=1)
                     ->  ProjectSet  (cost=0.00..0.00 rows=1 width=8) (actual time=0.263..6964.927 rows=100000 loops=1)
                           ->  Result  (cost=0.00..0.00 rows=1 width=1) (actual time=0.001..0.001 rows=1 loops=1)
 Planning Time: 40.596 ms
   (slice0)    Executor memory: 36K bytes avg x 3 workers, 36K bytes max (seg0).
   (slice1)    Executor memory: 61K bytes (entry db).
 Memory used:  128000kB
 Optimizer: Pivotal Optimizer (GPORCA)
 Execution Time: 7205.448 ms
(13 rows)

--after d874700
explain analyze insert into seq_test(j) select generate_series(1, 100000);
                                                             QUERY PLAN                                                             
------------------------------------------------------------------------------------------------------------------------------------
 Insert on seq_test  (cost=0.00..0.02 rows=1 width=8) (never executed)
   ->  Result  (cost=0.00..0.00 rows=1 width=12) (actual time=88.675..42203.682 rows=33462 loops=1)
         ->  Redistribute Motion 1:3  (slice1)  (cost=0.00..0.00 rows=1 width=8) (actual time=88.671..42198.652 rows=33462 loops=1)
               Hash Key: (int4(nextval('seq_test_i_seq'::regclass)))
               ->  Result  (cost=0.00..0.00 rows=1 width=8) (actual time=0.238..42013.601 rows=100000 loops=1)
                     ->  ProjectSet  (cost=0.00..0.00 rows=1 width=8) (actual time=0.237..41952.577 rows=100000 loops=1)
                           ->  Result  (cost=0.00..0.00 rows=1 width=1) (actual time=0.001..0.002 rows=1 loops=1)
 Planning Time: 22.819 ms
   (slice0)    Executor memory: 36K bytes avg x 3 workers, 36K bytes max (seg0).
   (slice1)    Executor memory: 61K bytes (entry db).
 Memory used:  128000kB
 Optimizer: Pivotal Optimizer (GPORCA)
 Execution Time: 42262.197 ms
(13 rows)

--after current fix
explain analyze insert into seq_test(j) select generate_series(1, 100000);
                                                            QUERY PLAN                                                             
-----------------------------------------------------------------------------------------------------------------------------------
 Insert on seq_test  (cost=0.00..0.02 rows=1 width=8) (never executed)
   ->  Result  (cost=0.00..0.00 rows=1 width=12) (actual time=82.371..7071.345 rows=33462 loops=1)
         ->  Redistribute Motion 1:3  (slice1)  (cost=0.00..0.00 rows=1 width=8) (actual time=82.368..7066.024 rows=33462 loops=1)
               Hash Key: (int4(nextval('seq_test_i_seq'::regclass)))
               ->  Result  (cost=0.00..0.00 rows=1 width=8) (actual time=0.243..6930.528 rows=100000 loops=1)
                     ->  ProjectSet  (cost=0.00..0.00 rows=1 width=8) (actual time=0.241..6875.615 rows=100000 loops=1)
                           ->  Result  (cost=0.00..0.00 rows=1 width=1) (actual time=0.001..0.002 rows=1 loops=1)
 Planning Time: 22.157 ms
   (slice0)    Executor memory: 36K bytes avg x 3 workers, 36K bytes max (seg0).
   (slice1)    Executor memory: 61K bytes (entry db).
 Memory used:  128000kB
 Optimizer: Pivotal Optimizer (GPORCA)
 Execution Time: 7117.249 ms
(13 rows)

```
And the gap should be higher for higher amount of data.